改寫推薦顯示，公式未導入

+++ app/src/main/java/com/mis/ncyu/quickchoice/result_type.java
@@ -8,11 +8,11 @@ public class result_type {
     private int id;
     private String name;
     private String keyword;
-    private Integer key;
+    private Double key;
     private Boolean chk = Boolean.FALSE;
     public result_type(){}
     public void compute_key(String replace){
-           this.key =Integer.parseInt(this.keyword.replace(replace,""));
+           this.key =Double.parseDouble(this.keyword.replace(replace,""));
     }
 
     public Boolean getChk() {
@@ -27,7 +27,7 @@ public class result_type {
         return id;
     }
 
-    public Integer getKey() {
+    public Double getKey() {
         return key;
     }
 
@@ -43,7 +43,7 @@ public class result_type {
         this.id = id;
     }
 
-    public void setKey(Integer key) {
+    public void setKey(Double key) {
         this.key = key;
     }
